LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-aspect-ratio-img-row-005-expected.txt

The green boxes should all be 100x100.

Section B of flex base size + transferred size suggestion:

Section B of flex base size + transferred size suggestion, with box-sizing: border-box:

content size suggestion

content size suggestion


PASS img 1 
FAIL img 2 assert_equals: 
<img style="height: 150px; box-sizing: border-box;" src="support/200x200-green.png" data-expected-client-width="100" data-expected-client-height="100">
clientWidth expected 100 but got 150
PASS img 3 
PASS img 4 
